Functions

Documentation for the functions included in the package.

Feature Generation

Function and types for generating features

EDDP.CellFeature — Type

CellFeature

Collection of Feature specifications and cell.

EDDP.CellFeature — Method

CellFeature(elements;kwargs...)

Construct a CellFeature instance.

Args

elements: A vector of the elements to be included in the features.

p2: A sequence of the two-body polynomial powers ($p$).

p3: A sequence of the three-body polynomial powers ($p$).

q3: A sequence of the three-body polynomial powers ($q$).

rcut2: Cut off distance for two-body features.

rcut3: Cut off distance for three-body features.

f2: Distance function for two-body interactions.

f3: Distance function for three-body interactions.

g2: Gradient function for two-body interactions.

g3: Gradient function for three-body interactions.

geometry_sequence: Wether to covnert the p2, p3, q3 as geometry sequence.

EDDP.ThreeBodyFeature — Type

(f::ThreeBodyFeature)(out::Vector, rij, rik, rjk, si, sj, sk)

EDDP.ThreeBodyFeature — Type

(f::ThreeBodyFeature)(out::Vector, rij, rik, rjk)

Accumulate an existing feature vector

EDDP.ThreeBodyFeature — Type

ThreeBodyFeature{T, M} <: AbstractNBodyFeature

Type for constructing the feature vector of the three-body interactions.

EDDP.ThreeBodyFeature — Method

(f::ThreeBodyFeature)(rij) = f(zeros(nfeatures(f)), rij, rik, rjk)

EDDP.TwoBodyFeature — Type

(f::TwoBodyFeature)(out::AbstractMatrix, rij, iat, istart=1)

Accumulate an existing matrix of the feature vectors

Args:

out: Output matrix

rji: distance between two atoms

iat: starting index of the vector to be updated

istart: starting index of the vector to be updated

EDDP.TwoBodyFeature — Type

TwoBodyFeature{T, M} <: AbstractNBodyFeature

Type for constructing the feature vector of the two-body interactions.

Base.:+ — Method

+(a::CellFeature, b::CellFeature)

Combine two CellFeature objects together. The features are simply concatenated in this case.

Base.:== — Method

==(A::T, B::T) where {T<:Union{AbstractNBodyFeature, CellFeature}}

Check equality between two CellFeature / AbstractNBodyFeature objects.

EDDP.fast_pow — Method

Faster version of (^) by expanding more integer power into multiplications

EDDP.feature_names — Method

feature_names(features...)

Return the name for the features.

EDDP.feature_names — Method

feature_names(cf::CellFeature)

Return the name for the features.

EDDP.feature_size — Method

feature_size(cf::CellFeature)

Return size of the feature vector for each body-order of a CellFeature object.

EDDP.feature_vector! — Method

feature_vector!(fvecs, gvecs, features2, features3, cell;kwargs...)

Args:

nl: If passed, using it as an existing NeighbourList.

Compute the feature vector for a given set of two and three body interactions, This is an optimised version for feature generation, but does not compute the gradients.

Returns the feature vector and the core repulsion energy if any.

EDDP.feature_vector — Method

feature_vector(cf::CellFeature, cell::Cell; nmax=500, skin=1.0)

Return a matrix of vectors describing the environment of each atom.

EDDP.feature_vector2! — Method

feature_vector(features::Vector{T}, cell::Cell) where T

Compute the feature vector for a give set of two body interactions

EDDP.feature_vector3! — Method

feature_vector(features::Vector{T}, cell::Cell) where T

Compute the feature vector for each atom a give set of three body interactions

EDDP.fr — Method

fr(r, rcut)

Equation 7 in Pickard 2022 describing interactions with well-behaved cut offs:

\[f(r)= \begin{cases} 2(1 - r / r_{rc}) & r \leq r_{rc} \\ 0 & r > r_{rc} \end{cases}\]

EDDP.genp — Method

Generate p/q as a geometry series

EDDP.genp — Method

Generate p/q as geometry series

EDDP.gfr — Method

gfr(r, rcut)

Gradient of the Equation 7 in Pickard 2022 describing interactions with well-behaved cut offs:

\[g(r)= \begin{cases} -2 / r_{rc} & r \leq r_{rc} \\ 0 & r > r_{rc} \end{cases}\]

EDDP.nbodyfeatures — Method

nbodyfeatures(c::CellFeature, nbody)

Return the number of N-body features

EDDP.nfeatures — Method

features(c::CellFeature)

Return the total number of features elements in a CellFeature object.

EDDP.one_body_vectors! — Method

one_body_vectors!(v, cell::Cell, cf::CellFeature)

Construct one-body features for the structure. The one-body feature is essentially an one-hot encoding of the specie labels

EDDP.one_body_vectors — Method

one_body_vectors(cell::Cell, cf::CellFeature)

Construct one-body features for the structure. The one-body feature is essentially an one-hot encoding of the specie labels

EDDP.permequal — Method

permequal(A, i, j, k)

Check equivalence considering all permutations with the first element matched.

EDDP.permequal — Method

permequal(A, i, j)

Check equivalence considering all permutations with the first element matched.

EDDP.suggest_rcut — Method

suggest_rcut(features...; offset=1.0)

Get a suggested rcut for a collection of features.

EDDP.suggest_rcut — Method

suggest_rcut(cf::CellFeature; offset=1.0)

Get a suggested cut off radius for NN list for a CellFeature.

EDDP.three_body_feature_from_mapping — Function

three_body_feature_from_mapping(cell::Cell, p_mapping, q_mapping, rcut, func=fr)

Construct a vector containing the TwoBodyFeatures

EDDP.two_body_feature_from_mapping — Function

two_body_feature_from_mapping(cell::Cell, p_mapping, rcut, func=fr)

Construct a vector containing the TwoBodyFeatures

EDDP.withgradient! — Method

withgradient!(e::Matrix, g::Vector, f::TwoBodyFeature, rij, iat, istart)

Calculate d(f(r)^p) / dr for each feature as well as the feature vector vector.

Args:

e: The matrix storing the feature vectors of shape (nfe, nat)

g: The matrix storing the gradient feature vectors of shape (nfe,), for this particular pair of i and j.

EDDP.withgradient! — Method

Calculate df / drij, df /drik, df/drjk for each element of a ThreeBodyFeature

Args:

e: The matrix storing the feature vectors of shape (nfe, nat)

g: The matrix storing the gradient feature vectors of shape (3, nfe), against rij, rik and rjk for this particular combination of i, j, k.

EDDP.ForceBuffer — Type

ForceBuffer{T}

Buffer for storing forces and stress and support their calculations

EDDP.ForceBuffer — Method

Initialise a buffer for computing forces

EDDP._force_update! — Method

_force_update!(buffer::Array{T, 2}, gv, g)

Propagate chain rule to obtain the forces

EDDP._stress_update! — Method

_stress_update!(buffer::Array{T, 2}, sv, s) where {T}

Propagate chain rule to obtain the stress

EDDP.compute_fv! — Method

compute_fv!(fvecs, features2, features3, cell;nl,

Compute feature vectors only.

EDDP.compute_fv_gv! — Method

compute_fv_gv!(fvecs, gvecs, features2, features3, cell::Cell;nl=NeighbourList(cell, features[1].rcut))

Compute the feature vector for a given set of two and three body interactions, compute gradients as well. Optimised version with reduced computational cost....

EDDP.BodyEmbedding — Type

Embedding for a specific body order

EDDP.BodyEmbedding — Method

Apply the embedding for an matrix input, the columns are made of vectors to be embedded.

EDDP.BodyEmbedding — Method

Apply the embedding

EDDP.BodyEmbeddingGradient — Type

Compute the gradient of the embedding layer (W) x -> Embedding -> out

The embeding layer does is

\[out = x W\]

Operation is assumed to be in the "batch" mode, taking inputs of multiple columns

EDDP.CellEmbedding — Type

Embedding for a full CellFeature

EDDP.CellEmbedding — Method

Apply CellEmbedding to a full feature vector

This is written in such way to support autograd...

EDDP.CellEmbedding — Method

Apply CellEmbedding to a full feature vector

This is written in such way to support autograd...

EDDP.CellEmbeddingGradient — Type

Container to operate forward and backward passes

EDDP._apply_embedding — Method

Apply weighted averages to each column of a matrix, the weights are stored in a matrix as well.

\[O = F W\]

This is simply right multiply the weight matrix.

EDDP._apply_embedding_batch! — Method

Apply the embedding for an matrix input, the columns are made of vectors to be embedded.

EDDP._apply_embedding_batch — Method

Apply the embedding for an matrix input, the columns are made of vectors to be embedded.

EDDP.backprop! — Method

Perform backpropagation - this requires the forward pass to be run an upstream gradients computed

EDDP.forward! — Method

Forward pass

EDDP.forward! — Method

Forward pass

EDDP.num_each_feat_elements — Method

Number of elements for each feature

EDDP.num_embed — Method

Number of embedded vectors

EDDP.num_feat — Method

Number of feature sets to be compressed

ChainRulesCore.rrule — Method

Custom rule for the applyembedding_batch function

ChainRulesCore.rrule — Method

Custom rule for the applyembedding_cell function

EDDP.CoreReplusion — Type

CoreReplusion(rcut)

Construct a core repulsion with a specific cut off radius.

EDDP.core_repulsion — Method

Core repulsion energy

EDDP.gcore_repulsion — Method

Core repulsion gradient

Training

Routines for training models feature vectors

EDDP.FeatureContainer — Type

FeatureContainer{T,N}

Container for holding features from structures.

EDDP.FeatureContainer — Method

FeatureContainer(sc::StructureContainer, featurespec; nmax=500, kwargs...)

Get a feature container.

EDDP.StructureContainer — Type

StructureContainer{T}

Loader for storing structure data

EDDP.StructureContainer — Method

StructureContainer(paths::Vector)

Args:

- `energy_threshold`: structures with per-atom energy higher than this are excluded. 
  Relative to the median energy.

Base.split — Method

split(c::Container, n1, n2, ...;shuffle=true)

Split the container into multiple parts, each with N number of structures. By default, the output container will have their feature vectors standardized.

EDDP._idx_group_by_composition — Method

_idx_group_by_composition(structures)

Return the index of structures of each unique composition.

EDDP._select_per_atom_threshold — Method

Return index selected based on per-formula atomic energy

EDDP._split_vector — Method

Split a vector by integer numbres

EDDP.load_fc — Method

load_fc(fname)

Load FeatureContainer from a file.

EDDP.reconstruct_x! — Method

Recover training X inputs

EDDP.save_fc — Method

save_fc(fc, fname)

Save FeatureContainer into a file.

EDDP.standardize! — Method

standardize!(fc_train, fcs...)

Standardise multiple feature containers. Only the first argument is used for fitting.

EDDP.standardize! — Method

standardize!(fc::FeatureContainer)

Standardize the data in the FeatureContainer. The feature vectors are modified. Note that only the input features are fitted and scaled, the outputs are only fitted but the standardisation is not applied.

EDDP.standardize — Method

standardize(fc_train, fcs...)

Standardise multiple feature containers. Only the first argument is used for fitting.

EDDP.train_test_split — Method

tain_test_split(v::FeatureContainer; ratio_test=0.1, shuffle=true)

Split the training

EDDP.transform_x! — Method

Apply transformation for training X inputs

EDDP.TrainingResultsSummary — Method

TrainingResultsSummary(train, test ,valid)

Construct a TrainingResultsSummary object from TrainingResults for the train, test and validation sets.

Base.getindex — Method

Base.getindex(v::TrainingResults, idx::Union{UnitRange, Vector{T}}) where {T<: Int}

Allow slicing to work on TrainingResults.

EDDP.ae_per_atom — Method

Absolute per-atom error

EDDP.boltzmann — Function

boltzmann(x, kT, x0=0.)

Computes the boltzmann distribution: $e^{\frac{-(x - x0)}{kT}}$.

EDDP.create_ensemble — Method

create_ensemble(models::AbstractVector, x::AbstractVector, y::AbstractVector;

Create an EnsembleNNInterface from a vector of interfaces and x, y data for fitting.

EDDP.create_ensemble — Method

create_ensemble(all_models, fc_train::FeatureContainer, args...)

Create ensemble model from training data.

EDDP.ensemble_std — Method

ensemble_std(tr::TrainingResults{M, T};per_atom=true) where {M, T<:EnsembleNNInterface}

Return the standard deviation from the ensemble for each data point. Defaults to atomic energy.

EDDP.generate_chain — Method

Genreate a Chain based on a vector specifying the number of hidden nodes in each layer

EDDP.generate_f_g_optim — Method

generate_f_g_optim_batch(model, train, test)

Generate f, g!, view of the parameters and the callback function for NN training using Optim.

This is for 'batch' training where all of the data are included.

EDDP.generate_f_g_optim_alt — Method

generate_f_g_optim(model, train, test)

Generate f, g!, view of the parameters and the callback function for NN training using Optim.

EDDP.mae_per_atom — Method

mae_per_atom(itf::AbstractNNInterface, x, y, nat)

Return per-atom-mean absolute error.

EDDP.nnls_weights — Method

nnls_weights(models, x, y)

Compute the weights for an ensemble of models using NNLS. Args: - models: aTuple/Vectorof models. -x: aVectorcontaining the features of each structure. -y: aVector` containing the total energy of each structure.

EDDP.r2score_each_comp — Method

r2score_each_comp(tr::TrainingResults)

Compute the R2 score for each composition separately as it does not make sense to compute using the full dataset containing different compositions.

See also: https://scikit-learn.org/stable/modules/model_evaluation.html#r2-score

EDDP.rmse_per_atom — Method

rmse_per_atom(itf::AbstractNNInterface, x, y, nat)

Return per-atom root-mean square error.

EDDP.spearman — Method

spearman(tr::TrainingResults)

Compute the average spearman score for each composition.

EDDP.spearman_each_comp — Method

spearman_each_comp(tr::TrainingResults)

Return unique reduced formula and their spearman scores.

EDDP.train_lm! — Method

Perform training for the given TrainingConfig

EDDP.worker_train_one — Method

worker_train_one(model, x, y, jobs_channel, results_channel;kwargs...)

Train one model and put the results into a channel

EDDP.@_itf_per_atom_wrap — Macro

Allow func(itf, fc) signature to be used....

EDDP.ATWA! — Method

Implement the output of

$ A^T W A $

EDDP.ATWA_DIAG! — Method

Compute the output of

$ diag(A^T W A) $

EDDP.earlystopcheck — Method

Signal stop if the min is more than n cycles away

EDDP._chunk_ranges — Method

_chunk_ranges(ndata; ngps=1)

Divide a range into multiple chunks such that each thread receives up to ngps chunks in total.

EDDP._collect_gradient! — Method

_collect_gradient!(gvec, g::CellEmbeddingGradient, i)

Collect gradient from a CellEmbeddingGradient layer

EDDP._collect_gradient! — Method

_collect_gradient!(gvec, g::DenseGradient, i)

Collect gradient from a DenseGradient layer

EDDP.atomic_rmse — Method

atomic_rmse(f, x, y, yt)

Compute per atom based RMSE

Args:

f: prediction function

x: input data

y: reference data

yt: Normalisation transformation originally applied to obtain y.

EDDP.collect_gradients! — Method

collect_gradients!(gvec::AbstractVector, gbuff::ChainGradients)

Collect the gradients after back-propagration into a vector

EDDP.compute_jacobian_backprop! — Method

compute_jacobian_backprop!(jacmat, gbuffs, model, data::AbstractVector)

Compute jacobian of the NN weights/bias using back-propagation.

EDDP.default_rng_value — Method

default_rng_value()

Create an instance of the default RNG depending on Julia's version.

Julia version is < 1.7: Random.GLOBAL_RNG

Julia version is >= 1.7: Random.default_rng()

EDDP.setup_fg_backprop — Method

setup_fg_backprop(model, data::AbstractVector, y)

Get a function for computing both the objective function and the jacobian matrix for the mean atomic energy.

EDDP.setup_fj — Function

setup_fj(model::AbstractNNInterface, data::AbstractVector, y)

Setup the function returning the residual and the jacobian matrix.

Model interfaces

EDDP.backward! — Function

Do the backward step

EDDP.flux_mlp_model — Method

Obtain an baseline MLP model

EDDP.forward! — Function

Do the forward step

EDDP.gradinp! — Function

Return the gradient of the total energy against the inputs.

EDDP.gradparam! — Function

Return a vector of the gradients of the total energy against the parameters.

EDDP.load_from_jld2 — Method

load_from_jld2(str::AbstractString, t::Type{<:AbstractNNInterface})

Load saved interface from JLD2 file.

EDDP.load_from_jld2 — Method

load_from_jld2(str::AbstractString)

Load saved interface from JLD2 file.

EDDP.load_from_jld2 — Method

load_from_jld2(f)

Load from JLD2 file/JLD2 group.

EDDP.nparams — Function

Return the number of parameters

EDDP.paramvector — Function

Return a vector containing the parameters

EDDP.paramvector! — Function

Return a vector containing the parameters

EDDP.save_as_jld2 — Method

save_as_jld2(str::AbstractString, obj::AbstractNNInterface)

Save an trained interface object as JLD2.

EDDP.setparamvector! — Function

Set the parameters with a vector

EDDP.transformed_inp — Method

Obtain an transformed the input array

EDDP.load_from_jld2 — Method

load_from_jld2(f, obj::LinearInterface)

Load from JLD2 file/JLD2 group.

EDDP.save_as_jld2 — Method

save_as_jld2(f, obj::LinearInterface)

Save the interface into an opened JLD2 file/JLD2 group.

EDDP.FluxInterface — Type

Standard Flux interface wrapping a model object.

This interface uses Zygote for AD. Performance is not ideal for training using the LM method due to the accumulated overheads of calling the gradient function. It should be OK to use for inference (but still inferior compared to ManualFluxBackPropInterface).

EDDP.backward! — Method

Run the backward step - this creates the pull back functions

EDDP.DenseGradient — Type

Buffer for storing gradients of a dense network

y = σ.(W * x .+ b)

EDDP.DenseGradient — Method

DenseGradient(dense::Dense, gσ, n)

Buffer for storing gradients of a dense network

y = σ.(W * x .+ b)

Args:

gσ: the function that computes the local gradient of the activation function. It must take two arguments x and fx, where fx = σ(x).

n: size of the batch.

out: Matrix of the output after this layer

EDDP._get_or_create_chaingradients — Method

Select the ChainGradients with the right size

EDDP.backprop! — Method

backprop!(dg::DenseGradient, d::Dense;weight_and_bias=true)

Compute the gradients of a dense network based on back-propagation.

Args:

weight_and_bias: Update the gradients for weight and bias. Defaults to true.

EDDP.backward! — Method

backward!(chaing::ChainGradients, chain::Chain;gu=one(eltype(chain.layers[1].weight)))

Back-propagate the gradients after a forward pass. Args:

gu: is the upstream gradient for the loss of the entire chain. The default is equivalent to: loss = sum(chain(x)), which is a matrix of 1 in the same shape of the output matrix.

EDDP.forward! — Method

forward!(chaing::ChainGradients, chain::Chain, x;copy=false)

Do a forward pass compute the intermediate quantities for each layer

EDDP.forward! — Method

Internal per-layer forward pass

EDDP.gradinp! — Method

Return the gradient of the input matrix $X$ against the sum of the output $sum(G(X))$.

EDDP.load_from_jld2 — Method

load_from_jld2(f, obj::ManualFluxBackPropInterface)

Load from JLD2 file/JLD2 group.

EDDP.paramvector! — Method

paramvector!(vec, model)

Return a vector containing all parameters of a model concatenated as a vector

EDDP.save_as_jld2 — Method

save_as_jld2(f, obj::ManualFluxBackPropInterface)

Save the interface into an opened JLD2 file/JLD2 group.

EDDP.update_param! — Method

update_param!(model, param)

Update the parameters of a model from a given vector

EDDP.EnsembleNNInterface — Type

EnsembleNNInterface{T<:Tuple}

Ensemble of many models

EDDP.load_from_jld2 — Method

load_from_jld2(f::Union{<:JLD2.Group, <:JLD2.JLDFile}, ::Type{<:EnsembleNNInterface})

Load from JLD2 file or JLD2 group.

Using potentials

Routines for using the trained potentials for energy/forces/stress calculations.

EDDP.VariableCellCalc — Type

Filter for including lattice vectors in the optimisation

Reference: E. B. Tadmor, G. S. Smith, N. Bernstein, and E. Kaxiras, Phys. Rev. B 59, 235 (1999)

Base on ase.constraints.UnitCellFilter

CellBase.get_positions — Method

get_positions(cf::VariableCellCalc)

Composed positions vectors including cell defromations

CellBase.set_positions! — Method

Update the positions passing through the filter

EDDP._calculate_two_pass! — Method

Do a two-pass approach - first pass get the feature vectors and the second pass will obtain the gradients

EDDP._get_forces_and_stress — Method

_get_forces_and_stress(cf::VariableCellCalc;rebuild_nl)

Construct force and stress for the filter object

EDDP._rebuild_on_demand — Method

Trigger rebuild of the NeighbourList

EDDP.apply_chainrule! — Method

Apply Chain rule to compute forces

EDDP.check_global_minsep — Method

 check_global_minsep(nl::NeighbourList, threshold)

Check if there are atoms that are too close to each other.

EDDP.copycell! — Method

Copy the lattice and positions from one cell to the other

EDDP.get_energy — Method

get_energy(cell::Cell, cf::CellFeature, itf::AbstractNNInterface;kwargs...)

Convenient method for calling get_energy with a Calculator constructed ad-hoc.

EDDP.get_energy — Method

get_energy(calc::NNCalc; forces=false, rebuild_nl=true)

Return the total energy of the calculator.

EDDP.get_energy_std — Method

get_energy_std(cell::Cell, cf::CellFeature, itf::AbstractNNInterface;kwargs...)

Convenient method for calling getenergystd with a Calculator constructed ad-hoc.

EDDP.get_energy_std — Method

Return standard deviation of the predicted total energy Note: must be run after a energy call!

EDDP.get_enthalpy — Method

get_enthalpy(cell::Cell, cf::CellFeature, itf::AbstractNNInterface;kwargs...)

Convenient method for calling get_enthalpy with a Calculator constructed ad-hoc.

EDDP.get_forces — Method

get_forces(cell::Cell, cf::CellFeature, itf::AbstractNNInterface;kwargs...)

Convenient method for calling get_forces with a Calculator constructed ad-hoc.

EDDP.get_forces — Method

Forces including the stress contributions that is consistent with the augmented positions

EDDP.get_pressure — Method

get_pressure(cell::Cell, cf::CellFeature, itf::AbstractNNInterface;kwargs...)

Convenient method for calling get_pressure with a Calculator constructed ad-hoc.

EDDP.get_pressure — Method

get_pressure(calc::VariableCellCalc)

Return the total pressure with the external pressure subtracted.

EDDP.get_pressure_gpa — Method

get_pressure_gpa(vc::AbstractCalc)

Return pressure in unit of GPa.

EDDP.get_stress — Method

get_stress(vc::VariableCellCalc;kwargs...)

Return the effective stress of the VaraibleCellCalc (including the external pressure)

EDDP.is_equal — Method

is_equal(cell_a, cell_b)

Check if two cells are equal to each other

EDDP.update_feature_vector! — Method

update_feature_vector!(calc::NNCalc; rebuild_nl=true, gradients=true, global_minsep=0.01, maxvol=100, gv_offset=0)

Returns the updated the feature vectors after atomic displacements

EDDP.opt_tpsd — Method

opt_tpsd(vc::AbstractCalc; f_tol=1e-3, s_tol=1e-3, e_tol=1e-5, 
              itermax=1000, trace=false, α_tol=1e-10, callback=nothing, 
              trajectory=nothing)

Optimise a structure (calculation) using TPSD. Reference: - Barzilai, J. and Borwein, J.M., 1988. Two-point step size gradient methods. IMA journal of numerical analysis, 8(1), pp.141-148

Miscellaneous tools

Support utilities and tools.

CellBase.write_res — Method

write_res(path, vc::VariableCellCalc;symprec=1e-2, label="EDDP")

Write structure in VariableCellCalc as SHELX file.

EDDP.is_unique_fvec — Method

Check if a feature vector already present in an array of vectors

EDDP.stem — Method

Return the stea part of a file name

EDDP.update_metadata! — Method

update_metadata!(vc::AbstractCalc, label;symprec=1e-2)

Update the metadata attached to a Cell` object

EDDP.AbstractRecord — Type

Abstract record that can be used to construct a phase diagram

EDDP.PhaseDiagram — Method

PhaseDiagram(sc::StructureContainer)

Construct a PhaseDiagram object from a StructureContainer.

EDDP.PhaseDiagram — Method

PhaseDiagram(records::Vector{T})

Construct a PhaseDiagram from a vector of AbstractRecord. Args:

elements: A vector of symbols for enforcing a specific ordering of the elements.

EDDP.Simplex — Type

Type that represents a simplex in N-dimensional space

EDDP.Simplex — Method

Simplex(coords)

Construct a simplex from a matrix of the coordinates

EDDP._formation_energies — Method

_formation_energies(records, elemental_forms)

Return the formation energy per atom.

EDDP.bary_coords — Method

bary_coords(s::Simplex, point::Vector)

Obtain barycentric coordinates for a given point

EDDP.contains_point — Function

contains_point(s::Simplex, point::Vector, tol=1e-8)

Check if a Simplex contains the given point.

EDDP.coords_from_bary — Method

Convert barycentric coordinate to normal point

EDDP.get_2d_plot_data — Method

Return data needed for constructing a 2D convex hull diagram

EDDP.get_composition_coord — Method

get_composition_coord(comp, elements)

Return the composition coordinates

EDDP.get_coord — Method

get_coord(comp, elements)

Return the composition coordination for a given composition.

EDDP.get_decomposition — Method

get_decomposition(phased, record)

Return the decomposition of a record

EDDP.get_ternary_hulldata — Method

get_ternary_hulldata(phased::PhaseDiagram)

Return the ternary coordinates for all stable and non stable phases

EDDP._get_rel_target_and_pred — Method

Return relative per-atom energies

EDDP.enthalpyandvolume — Function

enthalpyandvolume(args...;kwargs...)

Plots the training/predicted enthalpy per atom verses the volume per atom. If the first argument is a TrainingResults then the predicted enthalpies will be used instead.

EDDP.inoutsample — Function

inoutsample(builder::Builder, i::Int)

Return plots for out-of-sample prediction analysis. This function loads the model trained up to $i$ iteration and apply it to the $i + 1$ iteration. The prediction results is hence effectively "out-of-sample".

Ideally, we want the out-of-sample to behave like the in-sample results. Unless otherwise stateed, latter refers to the training data and the model at the $i$ iteration.

EDDP.relativeabsoluteerror — Function

relativeabsoluteerror(t::TrainingResults, args...;kwargs...)

Plot the absolute error per atom against the relative energy (per atom). This can be useful to check that the those with lower energy has smaller errors.

EDDP.write_eddpf90_features — Method

Write the feature vectors

Automated potentials building

Function for providing automatic iterative potential building via random structure structure searching.

EDDP.Builder — Type

Builder(str::AbstractString="link.toml")

Load the builder from a YAML file. The file contains nested key-values pairs similar to the constructors of the types.

state:
    seedfile : "test.cell"

trainer:
    type : "locallm"
    nmodels : 128
cf:
    elements : ["H", "O"]
    # Power of the polynomials as geometry sequence 
    p2       : [2, 10, 5] 
    geometry_sequence : true

cf_embedding:
    n : 3

EDDP.Builder — Type

The Builder stores states and options for training and using the potentials.

EDDP._call_pp3 — Method

_call_pp3(aname)

Call pp3 to calculate the energy/pressure of a seed file.

EDDP._disp_get_completed_jobs — Method

_disp_get_completed_jobs(project_name)

Get the number of completed jobs as well as the total number of jobs under a certain project. NOTE: this requires disp to be available in the commandline.

EDDP._generate_random_structures — Method

_generate_random_structures(bu::Builder, iter)

Generate random structure (if needed) as training data for the Builder.

EDDP._input_structure_dir — Method

Directory for input structures

EDDP._make_symbol_keys — Method

_make_symbol_keys(dict::Dict)

Swap the String keys of a dictonary into Symbol.

EDDP._output_structure_dir — Method

Directory for output structures after external calculations

EDDP._perform_training_external — Method

Carry out training and save the ensemble as a JLD2 archive.

EDDP._run_external — Method

Run external code for generating training data

EDDP._run_rss_link — Method

_run_rss_link()

Run the random search step as part of the link! iterative building protocol. This function is intented to be called as a separated Julia process.

Note

This function is meanted to be called with commandline arguments.

EDDP._set_iteration! — Method

_set_iteration!(builder::Builder)

Fastword to the iteration by checking existing ensemble files

EDDP.get_energy — Method

get_energy(cell::Cell, builder::Builder, gen::Int=_latest_ensemble_iteration(builder);kwargs...)

Convenient method for calling get_energy using a Builder object.

EDDP.get_energy_std — Method

get_energy_std(cell::Cell, builder::Builder, gen::Int=_latest_ensemble_iteration(builder);kwargs...)

Convenient method for calling getenergystd using a Builder object.

EDDP.get_enthalpy — Method

get_enthalpy(cell::Cell, builder::Builder, gen::Int=_latest_ensemble_iteration(builder);kwargs...)

Convenient method for calling get_enthalpy using a Builder object.

EDDP.get_forces — Method

get_forces(cell::Cell, builder::Builder, gen::Int=_latest_ensemble_iteration(builder);kwargs...)

Convenient method for calling get_forces using a Builder object.

EDDP.get_pressure — Method

get_pressure(cell::Cell, builder::Builder, gen::Int=_latest_ensemble_iteration(builder);kwargs...)

Convenient method for calling get_pressure using a Builder object.

EDDP.link! — Method

link!(builder::Builder)

Run automated iterative building cycles.

EDDP.link — Method

link()

Run link! from command line interface.

Example:

julia -e "Using EDDP;EDDP.link()" -- --file "link.toml" 

EDDP.load_features — Method

load_features(bu::Builder, iteration...)

Loading features for specific iterations.

EDDP.parse_disp_output — Method

parse_disp_output(json_string)

Parse the output of disp db summary

EDDP.run_disp_castep — Method

Run relaxation through DISP

EDDP.run_pp3 — Function

run_pp3(file, seedfile, outpath)

Use pp3 for single point calculations.

EDDP.run_pp3_many — Method

run_pp3_many(workdir, indir, outdir, seedfile; n_parallel=1, keep=false)

Use PP3 for singlepoint calculation - launch many calculations in parallel.

EDDP.run_rss — Function

run_rss(str::AbstractString)

Run random structure searching for a configuration file for the builder.

EDDP.run_rss — Method

run_rss(builder::Builder)

Run random structures search using trained ensembel model. The output files are in the search subfolder by default.

EDDP.save_builder — Method

save_builder(fname::AbstractString, builder)

Save a Builder to a file.

EDDP.step! — Method

Run a step step for the Builder

EDDP.summarise — Method

Summarise the status of the project

EDDP.create_ensemble — Function

create_ensemble(bu::Builder, tra::TrainingOption=bu.trainer;save_and_clean=false, kwargs...)

Create ensemble from resulted models. Optionally save the created ensemble model and clear the transient data.

EDDP.dataset_name — Function

dataset_name(bu::Builder, i=bu.state.iteration)

Full path to the dataset file.

EDDP.load_training_dataset — Function

load_training_dataset(bu::Builder, iter=bu.state.iteration;combined=false)

Load dataset saved for training

EDDP.num_existing_models — Function

num_existing_models(bu::Builder, tra::TrainingOption=bu.trainer)

Return the number of existing models in the training directory.

EDDP.run_trainer — Function

run_trainer(trainer::TrainingOption, builder::Builder)

Train the model and write the result to the disk as a JLD2 archive.

EDDP.run_trainer — Method

run_trainer()

Run training through a command line interface.

EDDP.write_dataset — Function

write_dataset(builder)

Write the dataset to the disk. These are large JLD2 archive containing the data for training, testing and validation

None-negative least square routines

Internal routines in the NNLS.jl package embedded in the package.

EDDP.NNLS — Module

NNLS copied from NNLS.jl

Remove the code for the solver interface for reduced dependencies. Original code base: https://github.com/rdeits/NNLS.jl

EDDP.NNLS.QP — Type

QP{T}

Structure describing the QP: Minimize $\frac{1}{2} z' Q z + c' z$ Subject to $G z \leq g$

EDDP.NNLS.QPWorkspace — Method

QPWorkspace(qp::QP)

Construct a workspace and load problem data for the QP Minimize $\frac{1}{2} z' Q z + c' z$ Subject to $G z \leq g$

EDDP.NNLS.UnsafeVectorView — Type

UnsafeVectorView

Views in Julia still allocate some memory (since they need to keep a reference to the original array). This type allocates no memory and does no bounds checking. Use it with caution.

EDDP.NNLS.apply_householder! — Method

apply_householder!(u::AbstractVector{T}, up::T, c::AbstractVector{T}) where {T}

CONSTRUCTION AND/OR APPLICATION OF A SINGLE HOUSEHOLDER TRANSFORMATION.. Q = I + U*(U**T)/B The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 12, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM.

EDDP.NNLS.check_optimality_conditions — Method

check_optimality_conditions(qp::QP, z, λ)

Provies a scalar indicator of the violation of the KKT optimality conditions for a given QP, primal solution $z$, and dual solution $λ$. A feasible & optimal solution should return a value close to 0, while an infeasible or suboptimal solution should return a value greater than zero.

EDDP.NNLS.construct_householder! — Method

construct_householder!(u::AbstractVector{T}, up::T)::T where {T}

CONSTRUCTION AND/OR APPLICATION OF A SINGLE HOUSEHOLDER TRANSFORMATION.. Q = I + U*(U**T)/B The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 12, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM.

EDDP.NNLS.dual_infeasibility — Method

dual_infeasibility(qp::QP, λ)

Measure of primal infeasibility of the solution to the QP Minimize $\frac{1}{2} z' Q z + c' z$ Subject to $G z \leq g` where$z$is the primal solution and$λ`` is the dual solution. For a dual-feasible solution, this will return a value <= 0.

EDDP.NNLS.fastview — Method

fastview(parent::AbstractArray, start_ind::Integer, len::Integer) =

Fallback for non-contiguous arrays, for which UnsafeVectorView does not make sense.

EDDP.NNLS.fastview — Method

fastview(parent::Array{T}, start_ind::Integer, len::Integer) where {T}

UnsafeVectorView only works for isbitstype types. For other types, we're already allocating lots of memory elsewhere, so creating a new View is fine. This function looks type-unstable, but the isbitstype(T) test can be evaluated by the compiler, so the result is actually type-stable.

EDDP.NNLS.load! — Method

load!(work::QPWorkspace, Q, c, G, g)

Load problem data for the QP Minimize $\frac{1}{2} z' Q z + c' z$ Subject to $G z \leq g$

EDDP.NNLS.orthogonal_rotmat — Method

orthogonal_rotmat(a::T, b::T)::Tuple{T,T,T} where {T}

COMPUTE ORTHOGONAL ROTATION MATRIX.. The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 12, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM. COMPUTE.. MATRIX (C, S) SO THAT (C, S)(A) = (SQRT(A2+B2)) (-S,C) (-S,C)(B) ( 0 ) COMPUTE SIG = SQRT(A2+B2) SIG IS COMPUTED LAST TO ALLOW FOR THE POSSIBILITY THAT SIG MAY BE IN THE SAME LOCATION AS A OR B .

EDDP.NNLS.primal_infeasibility — Method

primal_infeasibility(qp::QP, z)

Measure of primal infeasibility of the solution to the QP Minimize $\frac{1}{2} z' Q z + c' z$ Subject to $G z \leq g` where$z$is the primal solution and$λ`` is the dual solution. For a primal-feasible solution, this will return a value <= 0.

EDDP.NNLS.slackness_violation — Method

slackness_violation(qp::QP, z, λ)

Measure of the violation of the complementary slackness condition for the QP. For an optimal solution, this should return a value near 0.

EDDP.NNLS.solve! — Method

solve!(work::NNLSWorkspace{T,TI}, max_iter::Integer=(3 * size(work.QA, 2)))

Algorithm NNLS: NONNEGATIVE LEAST SQUARES The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 15, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM. GIVEN AN M BY N MATRIX, A, AND AN M-VECTOR, B, COMPUTE AN N-VECTOR, X, THAT SOLVES THE LEAST SQUARES PROBLEM A * X = B SUBJECT TO X .GE. 0

EDDP.NNLS.solve! — Method

z, λ = solve!(work::QPWorkspace)

Solve the QP that was loaded into work using load!. Returns the primal solution $z$ and the dual solution $λ$.

EDDP.NNLS.solve_triangular_system! — Method

solve_triangular_system!(zz, A, idx, nsetp, jj)

The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 15, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM.

EDDP.NNLS.stationarity_violation — Method

stationarity_violation(qp::QP, z, λ)

Measure of the violation of the stationarity KKT condition for the QP. For an optimal solution, this should return a value near 0.